Refactored for Service folder in mautic_api_integration module, deleted entity from the mautic_segment module
