+++ web/modules/custom/mautic_api_integration/mautic_api_integration.services.yml
@@ -1,4 +1,4 @@
 services:
   mautic_api_integration.client:
-    class: Drupal\mautic_api_integration\MauticApiClient
+    class: Drupal\mautic_api_integration\Service\MauticApiClient
     arguments: ["@http_client", "@logger.factory", "@keyvalue"]
